+++ CHANGELOG.md
@@ -13,10 +13,13 @@
 - Renamed the `nothing` cardType to `normal`
 - Fixed card parsing from card codes
 # 0.1.3
+## General Changes
+- Made player count prompt effective when running from `run.py`, and ineffective when running from `Uno.py`
 ## Uno.py
 - Fixed `setup()` where it would give each player the amount of players times 7 cards
 <!-- - Fixed card matching on turn -->
 ### Game()
 - Fixed a syntax error in `deckCheck()`
+- Added more things in `_recite()`
 ### Card()
 - Made `Card().full` return a Title Case string