+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - Made player count prompt effective when running from `run.py`, and ineffective when running from `Uno.py`
 ## Uno.py
 - Fixed `setup()` where it would give each player the amount of players times 7 cards
-<!-- - Fixed card matching on turn -->
+- Began work on card matching on turn
 ### Game()
 - Fixed a syntax error in `deckCheck()`
 - Fixed `AttributeError: 'int' object has no attribute 'discard'` in line 191